adding support for conditional attributes
say you have a dropdown and inside that dropdown few items are selected and some are not. So you will need "selected" attribute on some elements which are selected in that dropdown.

after this change one can have a template with a conditional attribute like this
<option data-item="{binding}" conditional-attribute-selected = "{binding.isSelected ? 'true' : null}" value="{binding.value}">binding.label </option>

if isSelected is falsy then the attribute will not be added.